requirements.txt
mysql-connector-python==8.2.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.0
qrcode[pil]==7.4.2
email-validator==2.1.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
openpyxl==3.1.2
